Codigo/backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.3.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.2.2
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
setuptools==68.2.2
